feat: WIP started working on rendering in background thread
