+++ src/viewer/canvasscene/shaders/texture.wgsl
@@ -1,0 +1,16 @@
+@vertex
+fn vs_main(@builtin(vertex_index) index: u32) -> @builtin(position) vec4<f32> {
+    let uv = vec2f(vec2u((index << 1) & 2, index & 2));
+    return vec4f(uv * 2. - 1., 0., 1.);
+}
+
+@group(0) @binding(0)
+var tex: texture_2d<f32>;
+
+@group(0) @binding(1)
+var samp: sampler;
+
+@fragment
+fn fs_main(@location(1) uv: vec2<f32>) -> @location(0) vec4<f32> {
+    return textureSample(tex, samp, uv);
+}
